requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.27
openai==1.44.0
faiss-cpu==1.8.0
tiktoken==0.7.0
pypdf==4.3.1
python-dotenv==1.0.1
streamlit==1.38.0
pydantic==2.8.2
